batch 4 + E18-R3: H-11 (intdiv/hrtime cast), M4-8 finds two pg defects (V-42 permit leak on thread death, V-43 a pool per thread), research 28 (interposition binds inside libcurl, gate ~8 ns)
H-11 (sonnet, validated): the phpantom finding was intdiv(hrtime(true)) — hrtime is typed
int|float — not JSON-decoded fields as the BACKLOG entry said; a cast, phpstan L6 0 errors on
the example and php/grpc. BACKLOG corrected; note added that analyses need php/ignis.php beside
the stubs; H-12 for php/ignis.php's own findings.

M4-8 (bencher, re-run by main): bench/m4-pool-survives.sh. After a thread dies holding a
lease, created stays 2 but available stays 1 — the lease and its permit are orphaned in LEASES;
nothing ties a lease to its thread and unregister only fails HTTP responders (V-42). And a probe
of main's: --threads 3 gives pool_id 1, 2, 3 — every thread opens its own pool, so an app gets
threads × max connections (V-43). ADR-0015's 'process-wide' describes the Rust object, not the
deployment. Fixes M4-11 and M4-12 are main's, next, together with M4-1.

E18-R3 (main): docs/research/28. --export-dynamic-symbol exports read/poll/getaddrinfo from the
executable; poll called inside libcurl reached ours 8 times; Rust std keeps working with read
interposed via a direct syscall; gate ~8.3 ns per syscall on the non-fiber path; this libcurl
never calls getaddrinfo — acceptance (3) is per library, pending R1.

+++ examples/grpc_server.php
@@ -21,7 +21,7 @@ $methods = [
     '/ignis.Greeter/Countdown' => static function (Call $call): null {
         $n = (int) (Proto::decode($call->message())[1] ?? 3);
         for ($i = $n; $i >= 1; $i--) {
-            $call->send(Proto::encode([1 => $i, 2 => date('H:i:s') . '.' . sprintf('%03d', intdiv(hrtime(true), 1000000) % 1000)]));
+            $call->send(Proto::encode([1 => $i, 2 => date('H:i:s') . '.' . sprintf('%03d', intdiv((int) hrtime(true), 1000000) % 1000)]));
             Ignis\sleep(10);
         }
         return null;
